Add files via upload
cooling flow for an elliptical galaxy. the new paramteres suggested in the slides have been substituted in the code (mvir, rvir etc.). The central gas density is computed in order to have fb=0.16 as usual. Unfortunately, this makes now most of the ouput files filled with NaN. This does not happen if I choose smaller d0 (e-24) wrt to the d0= e-21 needed to obtain 0.16